Delete old inputs and begin new test structure
