--- tests/inputs/invalidInput1.txt
@@ -1,2 +1,0 @@
-int a = 2 + false;
-return a;
