--- tests/inputs/invalidInput2.txt
@@ -1,1 +1,0 @@
-const boolean a = true;a = 2; return a;
