--- tests/inputs/invalidInput3.txt
@@ -1,3 +1,0 @@
-int a = 2; 
-a = 2 + true;
-return a;
