--- tests/inputs/invalidInput4.txt
@@ -1,3 +1,0 @@
-boolean a = 1;
-a = 2 + false;
-a = false - true;
