--- tests/inputs/validInput1.txt
@@ -1,2 +1,0 @@
-int a = 2 + 3;
-return a;
