--- tests/inputs/validInput2.txt
@@ -1,1 +1,0 @@
-int a = 2 * 3;return a;
