--- tests/inputs/validInput3.txt
@@ -1,1 +1,0 @@
-int a = 2 - 3;return a;
