--- tests/inputs/validInput4.txt
@@ -1,1 +1,0 @@
-int a = 4 / 2;return a;
