--- tests/inputs/validInput7.txt
@@ -1,3 +1,0 @@
-boolean a = false + false;
-a = a + true;
-return a;
