--- tests/inputs/validInput8.txt
@@ -1,5 +1,0 @@
-boolean a = false;
-boolean b = true * true;
-a = b * true;
-b = b * false;
-return a;
