--- tests/inputs/validInput9.txt
@@ -1,3 +1,0 @@
-boolean a = true + false * true;
-int b = 0;
-b = 2 - 2 + 40;
